backend/requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
python-linkedin-v2==0.9.4
langchain==0.1.9
python-jose==3.3.0
flask-cors==4.0.0
gunicorn==21.2.0 
beautifulsoup4==4.12.2
lxml==4.9.3
openai>=1.0.0,<2.0.0